+++ subprojects/api/build.gradle
@@ -12,6 +12,7 @@ dependencies {
     //swagger
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.3.0'
 
+    implementation project(':gpt')
 }
 
 tasks.named('test') {
