+++ subprojects/api/src/main/java/com/poemfoot/api/ApiApplication.java
@@ -2,8 +2,10 @@ package com.poemfoot.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
+@ComponentScan(basePackages = {"com.poemfoot.api", "com.poemfoot.gpt"})
 public class ApiApplication {
 
     public static void main(String[] args) {
