+++ subprojects/api/src/main/java/com/poemfoot/api/domain/gpt/domain/BaseTime.java
@@ -1,4 +1,4 @@
-package com.poemfoot.gpt.domain;
+package com.poemfoot.api.domain.gpt.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
