+++ subprojects/api/src/main/java/com/poemfoot/api/domain/gpt/domain/GptAnswer.java
@@ -1,15 +1,10 @@
-package com.poemfoot.gpt.domain.answer;
+package com.poemfoot.api.domain.gpt.domain;
 
-import com.poemfoot.gpt.domain.BaseTime;
-import com.poemfoot.gpt.dto.response.chat.GptChatPoemResponse;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import java.util.List;
-import java.util.Objects;
-import java.util.stream.Collectors;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
