+++ subprojects/api/src/main/java/com/poemfoot/api/domain/gpt/repository/GptAnswerRepository.java
@@ -1,5 +1,6 @@
-package com.poemfoot.gpt.domain.answer;
+package com.poemfoot.api.domain.gpt.repository;
 
+import com.poemfoot.api.domain.gpt.domain.GptAnswer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GptAnswerRepository extends JpaRepository<GptAnswer,Long> {
