+++ subprojects/api/src/main/java/com/poemfoot/api/domain/gpt/repository/GptQuestionRepository.java
@@ -1,5 +1,6 @@
-package com.poemfoot.gpt.domain.question;
+package com.poemfoot.api.domain.gpt.repository;
 
+import com.poemfoot.api.domain.gpt.domain.GptQuestion;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
