--- subprojects/api/src/main/java/com/poemfoot/api/domain/hello/controller/HelloController.java
@@ -1,20 +1,0 @@
-package com.poemfoot.api.domain.hello.controller;
-
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@Tag(name = "Hello", description = "테스트")
-@RequestMapping("/hello")
-public class HelloController {
-
-    @Operation(summary = "Swagger 테스트")
-    @GetMapping
-    public ResponseEntity<String> hello() {
-        return ResponseEntity.ok("hello");
-    }
-}
