+++ subprojects/api/src/main/java/com/poemfoot/api/domain/member/repository/MemberRepository.java
@@ -1,4 +1,4 @@
-package com.poemfoot.api.domain.member.dao;
+package com.poemfoot.api.domain.member.repository;
 
 import com.poemfoot.api.domain.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
