--- subprojects/gpt/src/main/java/com/poemfoot/gpt/GptApplication.java
@@ -1,13 +1,0 @@
-package com.poemfoot.gpt;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class GptApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(GptApplication.class, args);
-    }
-
-}
