+++ subprojects/gpt/src/main/resources/application.yml
@@ -1,31 +1,2 @@
-spring:
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: ${MYSQL_URL}
-    username: ${MYSQL_USERNAME}
-    password: ${MYSQL_PASSWORD}
-  jpa:
-    hibernate:
-      ddl-auto: create
-    open-in-view: false
-    properties:
-      hibernate:
-        create_empty_composites:
-          enabled: true
-        format_sql: true
-        default_batch_fetch_size: 100
-
 gpt:
-  model: gpt-4-0125-preview
   token: ${GPT_TOKEN}
-
-springdoc:
-  packages-to-scan: com.poemfoot.gpt
-  default-consumes-media-type: application/json;charset=UTF-8
-  default-produces-media-type: application/json;charset=UTF-8
-  swagger-ui:
-    path: /gpt/swagger-ui.html
-    disable-swagger-default-url: true
-    display-request-duration: true
-    operations-sorter: alpha
-    tags-sorter: alpha
