+++ subprojects/gpt/src/test/java/com/poemfoot/gpt/service/GptPoemProviderTest.java
@@ -1,20 +1,17 @@
 package com.poemfoot.gpt.service;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 import com.poemfoot.gpt.domain.question.GptQuestionRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 
 @SpringBootTest
 @Transactional
-class GptPoemServiceTest {
+class GptPoemProviderTest {
 
     @Autowired
-    private GptPoemService gptPoemService;
+    private GptPoemProvider gptPoemProvider;
     @Autowired
     private GptQuestionRepository gptQuestionRepository;
